+++ src/test/kotlin/com/github/shmvanhouten/adventofcode2019/day07/Day07Test.kt
@@ -10,7 +10,7 @@ class Day07Test {
 
     val input =
         "3,8,1001,8,10,8,105,1,0,0,21,38,59,84,97,110,191,272,353,434,99999,3,9,1002,9,2,9,101,4,9,9,1002,9,2,9,4,9,99,3,9,102,5,9,9,1001,9,3,9,1002,9,5,9,101,5,9,9,4,9,99,3,9,102,5,9,9,101,5,9,9,1002,9,3,9,101,2,9,9,1002,9,4,9,4,9,99,3,9,101,3,9,9,1002,9,3,9,4,9,99,3,9,102,5,9,9,1001,9,3,9,4,9,99,3,9,101,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,2,9,4,9,3,9,102,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,1002,9,2,9,4,9,99,3,9,1002,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,101,1,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,101,1,9,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,1,9,4,9,3,9,1001,9,1,9,4,9,3,9,1001,9,2,9,4,9,99,3,9,1001,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,102,2,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,1001,9,2,9,4,9,3,9,1001,9,2,9,4,9,3,9,1001,9,1,9,4,9,3,9,1002,9,2,9,4,9,3,9,102,2,9,9,4,9,3,9,1002,9,2,9,4,9,99,3,9,101,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,102,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,101,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,102,2,9,9,4,9,3,9,1001,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,1002,9,2,9,4,9,99,3,9,1001,9,1,9,4,9,3,9,102,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,1001,9,2,9,4,9,3,9,101,1,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,1001,9,1,9,4,9,3,9,102,2,9,9,4,9,3,9,1001,9,2,9,4,9,3,9,101,1,9,9,4,9,99"
-            .split(',').map { it.toInt() }
+            .split(',').map { it.toInt().toLong() }
 
     @Nested
     inner class Part1 {
@@ -18,27 +18,27 @@ class Day07Test {
         @Test
         internal fun testInput() {
             val intCode = IntCode(3, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0)
-            val thrusterSettings = listOf(4, 3, 2, 1, 0)
-            assertThat(computeThrusterSignal(intCode, thrusterSettings), equalTo(43210))
+            val thrusterSettings = listOf(4L, 3L, 2L, 1L, 0L)
+            assertThat(computeThrusterSignal(intCode, thrusterSettings), equalTo(43210L))
         }
 
         @Test
         internal fun `testInput highest signal`() {
             val intCode = IntCode(3, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0)
-            assertThat(computeMaxThrusterSignal(intCode), equalTo(43210))
+            assertThat(computeMaxThrusterSignal(intCode), equalTo(43210L))
         }
 
         @Test
         internal fun testInput2() {
             val intCode = IntCode(3,23,3,24,1002,24,10,24,1002,23,-1,23,
                 101,5,23,23,1,24,23,23,4,23,99,0,0)
-            val thrusterSettings = listOf(0,1,2,3,4)
-            assertThat(computeThrusterSignal(intCode, thrusterSettings), equalTo(54321))
+            val thrusterSettings = listOf(0L,1L,2L,3L,4L)
+            assertThat(computeThrusterSignal(intCode, thrusterSettings), equalTo(54321L))
         }
 
         @Test
         internal fun part1() {
-            assertThat(computeMaxThrusterSignal(IntCode(input)), equalTo(338603))
+            assertThat(computeMaxThrusterSignal(IntCode(input)), equalTo(338603L))
         }
     }
 
@@ -48,9 +48,9 @@ class Day07Test {
         internal fun testInput() {
             val intCode = IntCode(3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,
                 27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5)
-            val thrusterSettings = listOf(9,8,7,6,5)
+            val thrusterSettings = listOf(9L,8L,7L,6L,5L)
 
-            assertThat(computeFeedbackedThrusterSignal(intCode, thrusterSettings), equalTo(139629729))
+            assertThat(computeFeedbackedThrusterSignal(intCode, thrusterSettings), equalTo(139629729L))
         }
 
         @Test
@@ -58,7 +58,7 @@ class Day07Test {
             val intCode = IntCode(3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,
                 27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5)
 
-            assertThat(maxFeedbackedThrusterSignal(intCode), equalTo(139629729))
+            assertThat(maxFeedbackedThrusterSignal(intCode), equalTo(139629729L))
         }
 
         @Test
@@ -67,12 +67,12 @@ class Day07Test {
                 -5,54,1105,1,12,1,53,54,53,1008,54,0,55,1001,55,1,55,2,53,55,53,4,
                 53,1001,56,-1,56,1005,56,6,99,0,0,0,0,10)
 
-            assertThat(maxFeedbackedThrusterSignal(intCode), equalTo(18216))
+            assertThat(maxFeedbackedThrusterSignal(intCode), equalTo(18216L))
         }
 
         @Test
         internal fun part2() {
-            assertThat(maxFeedbackedThrusterSignal(IntCode(input)), equalTo(63103596))
+            assertThat(maxFeedbackedThrusterSignal(IntCode(input)), equalTo(63103596L))
         }
     }
